+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,10 +3,12 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.FriendsStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,20 +36,46 @@ public class UserService {
     }
 
     public void addFriend(long userId, long friendId) {
-        friendsStorage.addFriend(userId, friendId);
+        if (checkIfUserExists(userId) && checkIfUserExists(friendId)) {
+            friendsStorage.addFriend(userId, friendId);
+        }
     }
 
     public void deleteFriend(long userId, long friendId) {
-        friendsStorage.deleteFriend(userId, friendId);
+        if (checkIfUserExists(userId) && checkIfUserExists(friendId) && checkForFriendship(userId, friendId)) {
+            friendsStorage.deleteFriend(userId, friendId);
+        }
     }
 
     public List<User> getFriends(long userId) {
-        return friendsStorage.getFriends(userId);
+        List<User> friends = new ArrayList<>();
+        for (Long id : friendsStorage.getFriendsIds(userId)) {
+            friends.add(userStorage.getUserById(id));
+        }
+        return friends;
     }
 
     public List<User> getCommonFriends(long userId, long otherUserId) {
         return getFriends(userId).stream()
                 .filter(x -> getFriends(otherUserId).contains(x))
                 .collect(Collectors.toList());
+    }
+
+    private boolean checkIfUserExists(long id) {
+        User user = userStorage.getUserById(id);
+        if (user != null) {
+            return true;
+        } else {
+            throw new NotFoundException("Пользователь " + id + " не найден.");
+        }
+    }
+
+    private boolean checkForFriendship(long userId, long friendId) {
+        List<Long> userFriendsIds = friendsStorage.getFriendsIds(userId);
+        if (!userFriendsIds.contains(friendId)) {
+            throw new NotFoundException("Пользователь " + friendId + " не найден в друзьях " +
+                    "у пользователя " + userId + ".");
+        }
+        return true;
     }
 }
